fix(files): render PDF preview pages lazily instead of holding every page bitmap
The files-screen PDF preview eagerly rendered up to 50 pages to
full-device-width ARGB_8888 bitmaps with no dimension cap and held them
all until the dialog closed — ~330 MB for a 50-page A4 document on a
1080p phone, ~1.85 GB on a 2560px tablet — and showed only a spinner
until every page finished. Pages beyond 50 were silently dropped.

Extract PdfDocumentHolder from the chat PdfViewer into core:ui and
rebuild the preview on it: pages render on demand as they scroll into
view, are recycled when they scroll out, and are dimension-capped at
2048x8192, so memory tracks the visible window (~2-4 pages) instead of
the whole document. The 50-page cap is gone — every page is reachable
and "Page N of M" reports the true count. The spinner now covers only
download+open, and the staging temp file is unlinked immediately after
the fd opens (no more orphaned cache files on process death).

The whole-list pinch-zoom, page labels, and error fallback card are
unchanged. iOS actual untouched.
